SOAP-397: fixed nullpointer if wrong download link provided

+++ src/main/java/com/eviware/downloader/Downloader.java
@@ -225,7 +225,7 @@ public class Downloader
 
 	public static void removeTempFile()
 	{
-		if( tempFile.exists() )
+		if( tempFile != null && tempFile.exists() )
 		{
 			tempFile.deleteOnExit();
 		}
